Refactor : Rename api.py to api_attraction.py.
